Checked condition to not delete a file if its getting processed

+++ routes/api/documents.js
@@ -121,6 +121,8 @@ router.delete("/delete/:id", authenticate, async (req, res) => {
 
     if (!document) throw new HTTPError(400, "File not found");
 
+    if (!document.processing) throw new HTTPError(400, "File is currently getting processed");
+
     deleteFile(req.params.id, (err) => {
       if (err) return res.status(500).json({ success: false, message: "Failed to delete file" });
